Coding-Projects/antigravity/newsletter-reader/requirements.txt
Flask>=3.0
Flask-SQLAlchemy>=3.1
google-api-python-client>=2.100
google-auth-httplib2>=0.2
google-auth-oauthlib>=1.2
google-generativeai>=0.8
python-dotenv>=1.0
beautifulsoup4>=4.12
